+++ CHANGELOG.md
@@ -10,9 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Detect greentext when it is not at the beginning of a line.
 
+### Fixed
+
+- Don't replace custom emojis in links.
+
 ### Security
 
 - Don't create greentext elements in non-text nodes.
+- Don't replace custom emoji shortcodes in non-text nodes.
 
 ## [3.20.0] - 2025-03-15